Use _WIN32, not __GNUC__, to detect Windows.
Compilers masquerading as GCC (i.e., not just GCC but also clang, icc,
mingw, etc.) define __GNUC__ but may be targeting Windows.
Furthermore, there may be compilers targeting non-Windows platforms
which do not define __GNUC__.  _WIN32 can be used reliably to check if
the target platform is Windows, whereas __GNUC__ should only be used to
check if the compiler is GCC (or something masquerading it which
accepts the same code).