Don't depend on unaligned loads by default.
Using memcpy is just as fast with almost all modern compilers but does
not rely on undefined behavior.  Compilers can (and do) assume that
input programs do not depend on undefined behavior and are therefore
free to use instructions which require that the input is aligned, even
on CPUs like x86 where most instructions allow unaligned access.

Clang and MSVC all generate basically the same code for most unaligned
accesses as they do for memcpy with a fixed size of 4 bytes, and have
done so for quite a while.  GCC has long done the same on x86, but did
not do so on ARM (and possibly other architectures) until version 6.
Intel is the only compiler I'm aware of which still generates
sub-optimal code for memcpy with the current version of the compiler
(16); the issue has been reported.